router: Require that `Recognize` implement `Clone`
In the proxy, it's trivial for `Recognize` to be cloneable. If `Router`
takes advantage of this fact, the scope of the router's mutex can be
made smaller.

Additionally, this narrows the scope of the `Inner` type to only handle
caching logic.

There are no functional changes, though the (unused) `Single` router has
been removed because it could not trivially implement `Clone`