bin/conduit: Extract CLI binaries into a tag-specific dir
bin/conduit extracts CLI binaries out of a docker image, but it's easy
for cached versions of the binary to hide updates.

This change causes extracted images to be placed in tag-specific
directories so that bin/conduit reliably uses the most up-to-date
version of the CLI.

bin/conduit deletes unused directories.