proxy: Prevent active services from being evicted
It's possible for there to be services that are not accessed by the
router but have active streams (e.g. downloads, long-lived streams).
It's not appropriate to evict services that are busy doing work.

This change adds a `Retain` trait to the router, encapsulating the
decision of whether an idle node may be evicted.