proxy: Require that HTTP bodies implement Debug
When debugging proxy issues, it's helpful to be able to log details
about a stream body. However, because we are generally generic over the
body type, and the `Body` trait does not require that implementors also
implement `Debug`, then we must make this requirement explicit for all
uses.

This change adds `Debug` bounds to all body types. There is no
functional change.