Proxy: Make the control plane completely optional.
Previously, the proxy required `CONDUIT_PROXY_CONTROL_URL` to be set,
bit it still worked fine even if it could not contact any server at
that address, modulo its inability to do Kubernetes service discovery.
However, the proxy would keep trying to connect over and over again,
so such a degenerate configuration didn't work so well. In particular,
the logs were very noisy.

With this change, `CONDUIT_PROXY_CONTROL_URL` is now optional. See the
doc comments in config.rs for the difference in semantics when it is
set vs. when it is unset.

Fixes the remaining part of #486.

Signed-off-by: Brian Smith <brian@briansmith.org>