wip: proxy: prototype splitting grpc client from h2 for testing
The destination discovery code is generic over an H2 service, meaning
that, to provide it with a stubbed implementation of the grpc service,
we must actually implement a full, networked h2 service.

This change demonstrates an approach for making the destination
discovery logic generic over the grpc service itself so that stubbed
implementations may more easily be provided for testing.