Set outbound capacity on proxy in Prometheus pod
The `conduit-proxy` injected into the control-plane's Prometheus pod
makes requests to every other mesh-enabled pod in the cluster. For a
large number of pods (>300), this triggered
`conduit_proxy router at capacity` errors from the proxy, as the default
outbound capacity is set at 100.

This change increases `CONDUIT_PROXY_OUTBOUND_ROUTER_CAPACITY` from 100
to 10000 on the `conduit-proxy` running in the Prometheus pod.

Relates to
https://github.com/runconduit/conduit/issues/1179#issuecomment-400800995

Signed-off-by: Andrew Seigner <siggy@buoyant.io>