Add flag that skips `dep ensure` to bin/fast-build (#1321)
* Add flag that skips `dep ensure` to bin/fast-build

bin/fast-build is supposed to be fast. `dep ensure -vendor-only` is too slow
to meet this goal. Add `LINKERD_SKIP_DEP` to allow skipping it. The default
behavior is kept as-is to reduce new users' confusion.

The difference in speed isn't too notable now because the bin/docker-build
step drowns out the win currently. But if/when the bin/docker-build step is
replaced, this matters a lot.

Signed-off-by: Brian Smith <brian@briansmith.org>