Make use of new src and dst metadata hydration (#1444)
Now that we have source metadata in tap events, we can display 
the pod name in the UI instead of the IP. I've also added a popover 
that shows deploy and pod info if we have it.

Also adds another table in the expanded row view to show all the 
metadata we have. This table probably won't stick around forever, 
but I'm just displaying all the data we have right now.