Move more info from the tap table into the expanded row (#1641)
Try to make the tap table easier to parse by moving some info into the expanded
row. You can also now click anywhere on the row to expand.

The mocks in #1629 have Authority, Path and Latency buried, but I figured they might be
useful to see in the top level, so they're here.